Ray_demo/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@d5894ab4992878c191ed4c9fdb53b35ad2c0e628
aiofiles==23.1.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bitsandbytes==0.41.0
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
cmake==3.27.0
coloredlogs==15.0.1
colorful==0.5.5
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
einops==0.6.1
embeddings==0.0.8
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
ffmpy==0.3.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
gpustat==1.1
gradio==3.39.0
gradio_client==0.3.0
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
InstructorEmbedding==1.0.1
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
langchain==0.0.247
langsmith==0.0.15
linkify-it-py==2.0.2
lit==16.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
nvidia-ml-py==12.535.77
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
opencensus==0.11.2
opencensus-context==0.1.3
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydantic_core==2.4.0
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.14.0
PyPika==0.48.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
ray==2.6.2
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.5.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboardX==2.6.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchdata==0.6.1
torchtext==0.15.2+cpu
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@4a564490e149620282b9e946027e851f5f93aeae
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
widgetsnbextension==4.0.8
xformers==0.0.20
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.16.2
